Fix algorithm Profile Edit/Delete

+++ application/controllers/admin/Profile.php
@@ -255,103 +255,114 @@ class Profile extends CI_Controller
 	{
 		if ($this->session->userdata('role') == 'admin') {
 
-    		// If there is no profile in this id, then redirect
-			$tot = $this->Model_profile->profile_check($id);
-			if(!$tot) {
+			$data['profile'] = $this->Model_profile->getData($id);
+
+			if ($this->session->userdata('full_name') == $data['profile']['full_name']) {
 				redirect(base_url().'admin/profile');
 				exit;
-			}
-
-			$header['setting'] = $this->Model_header->get_setting_data();
-			$data['error'] = '';
-			$data['success'] = '';
-			$error = '';
-
-
-			if(isset($_POST['form_update'])) 
-			{
-				$valid = 1;
-
-				$this->form_validation->set_rules('full_name', 'Full Name', 'trim|required');
-				$this->form_validation->set_rules('email', 'Email', 'trim|required');
-				$this->form_validation->set_rules('phone', 'Phone', 'trim|required');
-				$this->form_validation->set_rules('password', 'Password', 'trim|required');				
-				$this->form_validation->set_rules('re_password', 'Retype Password', 'trim|required|matches[password]');
-
-				if($this->form_validation->run() == FALSE) {
-					$valid = 0;
-					$error .= validation_errors();
-				}
-
-				$path = $_FILES['photo']['name'];
-				$path_tmp = $_FILES['photo']['tmp_name'];
-
-				if($path!='') {
-					$ext = pathinfo( $path, PATHINFO_EXTENSION );
-					$file_name = basename( $path, '.' . $ext );
-					$ext_check = $this->Model_header->extension_check_photo($ext);
-					if($ext_check == FALSE) {
-						$valid = 0;
-						$error .= 'Anda harus mengunggah file jpg, jpeg, gif atau png untuk foto unggulan<br>';
-					}
-				}
-
-				if($valid == 1) 
-				{
-					$data['profile'] = $this->Model_profile->getData($id);
-					
-					if($path == '') {
-						$form_data = array(
-							'full_name'	=> $_POST['full_name'],
-							'email' 	=> $_POST['email'],
-							'phone'     => $_POST['phone'],
-							'password'  => md5($_POST['password']),
-							'role'     	=> $_POST['role'],
-							'status'    => $_POST['status'],
-							'token'     => $_POST['token']
-						);
-						$this->Model_profile->update_user($id,$form_data);
-					}
-					else {
-						unlink('./public/uploads/'.$data['profile']['photo']);
-
-						$final_name = 'avatar-'.$id.'.'.$ext;
-						move_uploaded_file( $path_tmp, './public/uploads/'.$final_name );
-
-						$form_data = array(
-							'full_name'	=> $_POST['full_name'],
-							'email' 	=> $_POST['email'],
-							'phone'     => $_POST['phone'],
-							'password'  => md5($_POST['password']),
-							'photo'     => $final_name,
-							'role'     	=> $_POST['role'],
-							'status'    => $_POST['status'],
-							'token'     => $_POST['token']
-						);
-						$this->Model_profile->update_user($id,$form_data);
-					}
-
-					//Add Log User
-					helper_log("edit", '[EDIT] Data User: '.$_POST['full_name'].' diubah oleh '.$this->session->userdata('full_name'));
-
-					$data['success'] = 'profile telah berhasil diupdate';
-				}
-				else
-				{
-					$data['error'] = $error;
-				}
-
-				$data['profile'] = $this->Model_profile->getData($id);
-				
-				$this->load->view('admin/view_header',$header);
-				$this->load->view('admin/view_profile_edit',$data);
-				$this->load->view('admin/view_footer');
 
 			} else {
-				$data['profile'] = $this->Model_profile->getData($id);
-				$this->load->view('admin/view_header',$header);
-				$this->load->view('admin/view_profile_edit',$data);
-				$this->load->view('admin/view_footer');
+
+				// If there is no profile in this id, then redirect
+				$tot = $this->Model_profile->profile_check($id);
+				if(!$tot) {
+					redirect(base_url().'admin/profile');
+					exit;
+				}
+
+				$header['setting'] = $this->Model_header->get_setting_data();
+				$data['error'] = '';
+				$data['success'] = '';
+				$error = '';
+
+
+				if(isset($_POST['form_update'])) 
+				{
+					$valid = 1;
+
+					$this->form_validation->set_rules('full_name', 'Full Name', 'trim|required');
+					$this->form_validation->set_rules('email', 'Email', 'trim|required');
+					$this->form_validation->set_rules('phone', 'Phone', 'trim|required');
+					$this->form_validation->set_rules('password', 'Password', 'trim|required');				
+					$this->form_validation->set_rules('re_password', 'Retype Password', 'trim|required|matches[password]');
+
+					if($this->form_validation->run() == FALSE) {
+						$valid = 0;
+						$error .= validation_errors();
+					}
+
+					$path = $_FILES['photo']['name'];
+					$path_tmp = $_FILES['photo']['tmp_name'];
+
+					if($path!='') {
+						$ext = pathinfo( $path, PATHINFO_EXTENSION );
+						$file_name = basename( $path, '.' . $ext );
+						$ext_check = $this->Model_header->extension_check_photo($ext);
+						if($ext_check == FALSE) {
+							$valid = 0;
+							$error .= 'Anda harus mengunggah file jpg, jpeg, gif atau png untuk foto unggulan<br>';
+						}
+					}
+
+					if($valid == 1) 
+					{
+						$data['profile'] = $this->Model_profile->getData($id);
+
+						if($path == '') {
+							$form_data = array(
+								'full_name'	=> $_POST['full_name'],
+								'email' 	=> $_POST['email'],
+								'phone'     => $_POST['phone'],
+								'password'  => md5($_POST['password']),
+								'role'     	=> $_POST['role'],
+								'status'    => $_POST['status'],
+								'token'     => $_POST['token']
+							);
+							$this->Model_profile->update_user($id,$form_data);
+						}
+						else {
+							unlink('./public/uploads/'.$data['profile']['photo']);
+
+							$final_name = 'avatar-'.$id.'.'.$ext;
+							move_uploaded_file( $path_tmp, './public/uploads/'.$final_name );
+
+							$form_data = array(
+								'full_name'	=> $_POST['full_name'],
+								'email' 	=> $_POST['email'],
+								'phone'     => $_POST['phone'],
+								'password'  => md5($_POST['password']),
+								'photo'     => $final_name,
+								'role'     	=> $_POST['role'],
+								'status'    => $_POST['status'],
+								'token'     => $_POST['token']
+							);
+							$this->Model_profile->update_user($id,$form_data);
+						}
+
+					//Add Log User
+						helper_log("edit", '[EDIT] Data User: '.$_POST['full_name'].' diubah oleh '.$this->session->userdata('full_name'));
+
+						$data['success'] = 'profile telah berhasil diupdate';
+					}
+					else
+					{
+						$data['error'] = $error;
+					}
+
+					$data['profile'] = $this->Model_profile->getData($id);
+
+					$this->load->view('admin/view_header',$header);
+					$this->load->view('admin/view_profile_edit',$data);
+					$this->load->view('admin/view_footer');
+
+				} else {
+
+					$data['profile'] = $this->Model_profile->getData($id);
+					$this->load->view('admin/view_header',$header);
+					$this->load->view('admin/view_profile_edit',$data);
+					$this->load->view('admin/view_footer');
+				}
+
 			}
 
 		} else {
@@ -368,25 +379,35 @@ class Profile extends CI_Controller
 	{
 		if (($this->session->userdata('role') == 'admin')) {
 
-		// If there is no profile in this id, then redirect
-			$tot = $this->Model_profile->profile_check($id);
-			if(!$tot) {
-				redirect(base_url().'admin/profile');
-				exit;
-			}
-
 			$data['profile'] = $this->Model_profile->getData($id);
 
-			if($data['profile']) {
-				unlink('./public/uploads/'.$data['profile']['photo']);
+			if ($this->session->userdata('full_name') == $data['profile']['full_name']) {
+				redirect(base_url().'admin/profile');
+				exit;
+
+			} else {
+				// If there is no profile in this id, then redirect
+				$tot = $this->Model_profile->profile_check($id);
+				if(!$tot) {
+					redirect(base_url().'admin/profile');
+					exit;
+				}
+
+				$data['profile'] = $this->Model_profile->getData($id);
+
+				if($data['profile']) {
+					unlink('./public/uploads/'.$data['profile']['photo']);
+				}
+
+				$this->Model_profile->delete_user($id);
+
+				//Add Log User
+				helper_log("delete", '[HAPUS] Data User Id:'.$data['profile']['full_name'].' dihapus oleh '.$this->session->userdata('full_name'));
+
+				redirect(base_url().'admin/profile');
+
 			}
 
-			$this->Model_profile->delete_user($id);
-
-			//Add Log User
-			helper_log("delete", '[HAPUS] Data User Id:'.$data['profile']['full_name'].' dihapus oleh '.$this->session->userdata('full_name'));
-
-			redirect(base_url().'admin/profile');
 		} else {
 			if(!$this->session->userdata('id')) {
 				redirect(base_url().'admin/login');
